[LCR-208] fix: type 병합
