+++ src/types/HomeScreenType.ts
@@ -12,3 +12,12 @@ export type PopUpItem = {
   popupCloseDate: string;
   address: string;
 };
+
+export type HotPopUpItem = {
+  popupId: number;
+  popupName: string;
+  imageUrl: number; // 추후 string으로 변경
+  popupOpenDate: string;
+  popupCloseDate: string;
+  address: string;
+};
